+++ test/transform/resource/after-delombok/FieldNameConstantsBasic.java
@@ -6,7 +6,7 @@ public class FieldNameConstantsBasic {
 	String butPrintMePlease;
 	@java.lang.SuppressWarnings("all")
 	static final class Fields {
-		public static final java.lang.String butPrintMePlease = "butPrintMePlease";
 		public static final java.lang.String iAmADvdPlayer = "iAmADvdPlayer";
+		public static final java.lang.String butPrintMePlease = "butPrintMePlease";
 	}
 }