+++ test/transform/resource/after-ecj/FieldNameConstantsBasic.java
@@ -2,8 +2,8 @@ import lombok.experimental.FieldNameConstants;
 import lombok.AccessLevel;
 public @FieldNameConstants(level = AccessLevel.PACKAGE) class FieldNameConstantsBasic {
   static final @java.lang.SuppressWarnings("all") class Fields {
-    public static final java.lang.String butPrintMePlease = "butPrintMePlease";
     public static final java.lang.String iAmADvdPlayer = "iAmADvdPlayer";
+    public static final java.lang.String butPrintMePlease = "butPrintMePlease";
     <clinit>() {
     }
     private @java.lang.SuppressWarnings("all") Fields() {